[MAJ] Classes CSS delete, edit et syndication renomm?es en pbt-icon-delete, pbt-icon-edit et pbt-icon-syndication Pr?c?demment la classe css delete afficher automatiquement une confirmation de suppression, ce n'est plus le cas, maintenant il faut utiliser data-confirmation="delete-element" Si vous souhaitez int?grer un message de confirmation simplement de vos codes, vous pouvez utiliser data-confirmation="votre message" Les boutons pour faire du tri dans les tableaux sont renomm? en pbt-icon-table-sort-up et pbt-icon-table-sort-down Ajout des classes CSS pbt-icon-plus et pbt-icon-minus (d'autres viendront)
Je pense que maintenant nous avons quelque chose de beaucoup plus coh?rent ;)

+++ articles/templates/ArticlesFormFieldSelectSources.tpl
@@ -46,11 +46,11 @@ var ArticlesFormFieldSelectSources = new ArticlesFormFieldSelectSources();
 	<div style="float:left;width:84%;display:inline;padding-bottom:5px;color:#334455;font-weight:bold;"><span style="width:50%;margin:0 75px 0 40px;">${i18n('articles.form.source_name')}</span><span style="width:50%;">${i18n('articles.form.source_url')}</span></div>
 	<input type="text" name="field_name_${escape(ID)}_{fieldelements.ID}" id="field_name_${escape(ID)}_{fieldelements.ID}" value="{fieldelements.NAME}" size="30" class="text">
 	<input type="text" name="field_value_${escape(ID)}_{fieldelements.ID}" id="field_value_${escape(ID)}_{fieldelements.ID}" value="{fieldelements.VALUE}" size="30" class="text"/>
-	<a href="javascript:ArticlesFormFieldSelectSources.delete_field({fieldelements.ID});" id="delete_${escape(ID)}_{fieldelements.ID}" class="delete"></a>
+	<a href="javascript:ArticlesFormFieldSelectSources.delete_field({fieldelements.ID});" id="delete_${escape(ID)}_{fieldelements.ID}" class="pbt-icon-delete" data-confirmation="delete-element"></a>
 	<br id="br_${escape(ID)}_{fieldelements.ID}">
 # END fieldelements #
 </div>
-<img src="{PATH_TO_ROOT}/templates/{THEME}/images/admin/plus.png" id="add_${escape(ID)}" class="valign_middle" style="width:25px;"/>
+<i class="pbt-icon-plus" id="add_${escape(ID)}"></i>
 <script type="text/javascript">
 <!--
 Event.observe(window, 'load', function() {
