+++ news/templates/NewsDisplaySeveralNewsController.tpl
@@ -20,7 +20,7 @@
 <section>
 	<header>
 		<h1>
-			<a href="${relative_url(SyndicationUrlBuilder::rss('news'))}" class="syndication" title="${LangLoader::get_message('syndication', 'main')}"></a>
+			<a href="${relative_url(SyndicationUrlBuilder::rss('news'))}" class="pbt-icon-syndication" title="${LangLoader::get_message('syndication', 'main')}"></a>
 			{L_NEWS_TITLE}
 		</h1>
 	</header>
@@ -36,14 +36,14 @@
 		<article # IF C_NEWS_BLOCK_COLUMN # style="float:left;width:{COLUMN_WIDTH}%" # ENDIF # itemscope="itemscope" itemtype="http://schema.org/CreativeWork">
 			<header>
 				<h1>
-					<a href="{news.U_SYNDICATION}" class="syndication" title="${LangLoader::get_message('syndication', 'main')}"></a>
+					<a href="{news.U_SYNDICATION}" class="pbt-icon-syndication" title="${LangLoader::get_message('syndication', 'main')}"></a>
         			<a href="{news.U_LINK}"><span id="name" itemprop="name">{news.NAME}</span></a>
         			<span class="actions">
         				# IF news.C_EDIT #
-							<a href="{news.U_EDIT}" title="${LangLoader::get_message('edit', 'main')}" class="edit"></a>
+							<a href="{news.U_EDIT}" title="${LangLoader::get_message('edit', 'main')}" class="pbt-icon-edit"></a>
 						# ENDIF #
 						# IF news.C_DELETE #
-							<a href="{news.U_DELETE}" title="${LangLoader::get_message('delete', 'main')}" class="delete"></a>
+							<a href="{news.U_DELETE}" title="${LangLoader::get_message('delete', 'main')}" class="pbt-icon-delete" data-confirmation="delete-element"></a>
 						# ENDIF #
 					</span>
         		</h1>
