+++ templates/default/admin/member/AdminViewAllMembersController.tpl
@@ -7,32 +7,32 @@
 					<th>
 					</th>
 					<th>
-						<a href="{SORT_LOGIN_TOP}" class="sort-up"></a>
+						<a href="{SORT_LOGIN_TOP}" class="pbt-icon-table-sort-up"></a>
 						{L_LOGIN} 
-						<a href="{SORT_LOGIN_BOTTOM}" class="sort-down"></a>
+						<a href="{SORT_LOGIN_BOTTOM}" class="pbt-icon-table-sort-down"></a>
 					</th>
 					<th>
-						<a href="{SORT_LEVEL_TOP}" class="sort-up"></a>
+						<a href="{SORT_LEVEL_TOP}" class="pbt-icon-table-sort-up"></a>
 						{L_LEVEL}
-						<a href="{SORT_LEVEL_BOTTOM}" class="sort-down"></a>
+						<a href="{SORT_LEVEL_BOTTOM}" class="pbt-icon-table-sort-down"></a>
 					</th>
 					<th>
 						{L_MAIL}
 					</th>
 					<th>
-						<a href="{SORT_REGISTERED_TOP}" class="sort-up"></a>
+						<a href="{SORT_REGISTERED_TOP}" class="pbt-icon-table-sort-up"></a>
 						{L_REGISTERED}
-						<a href="{SORT_REGISTERED_BOTTOM}" class="sort-down"></a>
+						<a href="{SORT_REGISTERED_BOTTOM}" class="pbt-icon-table-sort-down"></a>
 					</th>
 					<th>
-						<a href="{SORT_LAST_CONNECT_TOP}" class="sort-up"></a>
+						<a href="{SORT_LAST_CONNECT_TOP}" class="pbt-icon-table-sort-up"></a>
 						{L_LAST_CONNECT}
-						<a href="{SORT_LAST_CONNECT_BOTTOM}" class="sort-down"></a>
+						<a href="{SORT_LAST_CONNECT_BOTTOM}" class="pbt-icon-table-sort-down"></a>
 					</th>
 					<th>
-						<a href="{SORT_APPROBATION_TOP}" class="sort-up"></a>
+						<a href="{SORT_APPROBATION_TOP}" class="pbt-icon-table-sort-up"></a>
 						{L_APPROBATION}
-						<a href="{SORT_APPROBATION_BOTTOM}" class="sort-down"></a>
+						<a href="{SORT_APPROBATION_BOTTOM}" class="pbt-icon-table-sort-down"></a>
 					</th>
 				</tr>
 			</thead>
@@ -49,8 +49,8 @@
 				# START member_list #
 				<tr> 
 					<td> 
-						<a href="{member_list.EDIT_LINK}" title="{L_UPDATE}" class="edit"></a>
-						<a href="{member_list.DELETE_LINK}" title="{L_DELETE}" class="delete"></a>
+						<a href="{member_list.EDIT_LINK}" title="{L_UPDATE}" class="pbt-icon-edit"></a>
+						<a href="{member_list.DELETE_LINK}" title="{L_DELETE}" class="pbt-icon-delete" data-confirmation="delete-element"></a>
 					</td>
 					<td>
 						<a href="{member_list.U_PROFILE}" class="{member_list.LEVEL_CLASS}" # IF member_list.C_GROUP_COLOR # style="color:{member_list.GROUP_COLOR}" # ENDIF #>{member_list.LOGIN}</a>
